+++ crab-core/src/types/null.rs
@@ -1,4 +1,4 @@
-use crate::{Object, RawObjectData};
+use crate::object::{Object, RawObjectData};
 
 /// This struct represents nothing as in the absence of a value
 #[derive(Debug, Clone)]
@@ -9,7 +9,7 @@ impl Object for Null {
         Box::new(self.clone())
     }
 
-    fn into_raw_object_data(&self) -> crate::RawObjectData {
+    fn into_raw_object_data(&self) -> RawObjectData {
         RawObjectData::new([])
     }
 }
